Immediately re-establish long poll on message post
That is to say:
- if the long poll (/msg-wait) connection breaks
- and the client side is waiting to retry after 10sec
- when a new message is posted successfully,

we assume the success of posting the new msg implies that the network
connection (or server) is back up, and we can now *immediately*
re-establish the long poll (/msg-wait).

As a result, this eliminates the up to 10 second wait to see your own
message you just posted.

This commit also adds an alert if posting a message fails.